Initial fix of all warnings when compiling 64 bit
All warnings were fixed when compiling on a 64 bit Linux machine.
This was done in an attempt to possibly find additional 64 bit
incompatibilities.

+++ centrallix/osdrivers/objdrv_datafile.c
@@ -1611,7 +1611,7 @@ dat_internal_OpenNode(pDatData context, pObject obj, char* filename, int mode, i
 		return NULL;
 		}
 	    for(i=0;i<=strlen(ptr);i++) dn->Ext[i] = toupper(ptr[i]);
-	    dn->Type = (int)xhLookup(&DAT_INF.TypesByExt, (void*)(dn->Ext));
+	    dn->Type = (intptr_t)xhLookup(&DAT_INF.TypesByExt, (void*)(dn->Ext));
 	    if (!dn->Type)
 	        {
 		mssError(1,"DAT","Unknown datafile type '%s'",ptr);
