+++ centrallix/osdrivers/objdrv_sysinfo.c
@@ -263,7 +263,7 @@ sysAddAttrib(pSysInfoData sid, char* attrname, int type)
 	    }
 
 	/** Add the type **/
-	xaAddItem(sid->AttrTypes, (void*)type);
+	xaAddItem(sid->AttrTypes, (void*)(intptr_t)type);
 
     return 0;
     }
@@ -495,7 +495,7 @@ sysLoadAttrsArray(pSysData inf)
 		    t = -1;
 		else
 		    t = inf->Sid->GetAttrTypeFn(inf->Sid->Context, (inf->Flags & SYS_F_SUBOBJ)?inf->Name:NULL, xa->Items[i]);
-		xaAddItem(txa, (void*)t);
+		xaAddItem(txa, (void*)(intptr_t)t);
 		}
 	    }
 	inf->AttrTypesArray = txa;
@@ -902,7 +902,7 @@ sysGetAttrType(void* inf_v, char* attrname, pObjTrxTree* oxt)
 	    {
 	    if (!strcmp(inf->AttrsArray->Items[i], attrname))
 		{
-		t = (int)(inf->AttrTypesArray->Items[i]);
+		t = (intptr_t)(inf->AttrTypesArray->Items[i]);
 		return t;
 		}
 	    }
@@ -948,7 +948,7 @@ sysGetAttrValue(void* inf_v, char* attrname, int datatype, pObjData val, pObjTrx
 	    {
 	    if (!strcmp(inf->AttrsArray->Items[i], attrname))
 		{
-		t = (int)(inf->AttrTypesArray->Items[i]);
+		t = (intptr_t)(inf->AttrTypesArray->Items[i]);
 		if (t != datatype)
 		    {
 		    mssError(1,"SYS","Type mismatch requesting attribute '%s'", attrname);
